Add day4.ts and deal with duplicate def errors
Had to remove webpack (replaced it with Rake) as the ts-loader would try
and run tsc with all the ts files in the repo at the same time (setting
files value in tsconfigs didn't help) which caused lots of duplicate
definition errors.

The multiple tsconfig.json files are so VSCode knows that they are
different projects and doesn't give errors on duplicate definitions.